refactor(telemetry): move models to correct packages, IO dispatcher for disk ops
- StreamSignal → api/model/telemetry/ (data models package)
- StreamTelemetryConfig → api/model/config/ (alongside other configs)
- drain() is now suspend, disk reads on Dispatchers.IO
- spillToDisk launches on Dispatchers.IO via scope
- cleanStaleVersions runs on Dispatchers.IO at init
- StreamTelemetryImpl and StreamTelemetryScopeImpl accept CoroutineScope
