+++ stream-android-core/src/main/java/io/getstream/android/core/api/model/telemetry/StreamSignal.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.getstream.android.core.api.telemetry
+package io.getstream.android.core.api.model.telemetry
 
 import io.getstream.android.core.annotations.StreamInternalApi
